refactor(matrix): split matrix types module [OPE-425]
